Fix bundled Shell::Command to ignore already removed paths in rm_rf()

+++ ext/File__Tools/lib/Shell/Command.pm
@@ -24,6 +24,7 @@ sub rm_f(*@files) is export {
 
 sub rm_rf(*@files) is export {
     for @files -> $path {
+        next unless $path.IO.e;
         for find(dir => $path).map({ .Str }).reverse -> $f {
             $f.IO.d ?? rmdir($f) !! unlink($f);
         }
